chore: add home and login pages;
