+++ src/pages/Login/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Login'
+export { default } from './Login.container'
