+++ src/shared/styles/theme.scss
@@ -10,7 +10,7 @@
 $theme-default: (--default: $color-mine-shaft,
     --primary: $color-java,
     --secondary: $color-java,
-    --background: $color-black-squeeze,
+    --background: $color-java,
 );
 
 $theme-dark: (--default: $color-mine-shaft,
